Wrapper to prepare raw data

This is a wrapper function that returns only the desired columns, adds variables that can be derived to the data, and adds standardized scores.

Usage

prepare_data(
  dat,
  selected_cols = c("RACE", "CDRGLOB", "MOCATOTS", "MOCBTOTS", "TRAILA", "TRAILARR",
    "TRAILALI", "OTRAILA", "OTRLARR", "DIGFORCT", "DIGFORSL", "DIGBACCT", "DIGBACLS",
    "WAIS", "MINTTOTS", "ANIMALS", "VEG", "UDSVERTN", "UDSVERFC", "UDSVERLC", "UDSBENTC",
    "UDSBENTD", "CRAFTVRS", "CRAFTURS", "CRAFTDVR", "CRAFTDRE", "REY1REC", "REY2REC",
    "REY3REC", "REY4REC", "REY5REC", "REY6REC", "REYDREC", "REYTCOR", "TRAILB",
    "TRAILBLI", "TRAILBRR", "MOCACLOC", "MOCACLOH", "MOCACLON", "OTRAILB", "OTRLBRR",
    "OTRLBLI", "NACCGDS", "CDRSUM", 
     "UDSBENRS", "NACCID", "SEX", "EDUC", "HANDED",
    "NACCAGE", "BIRTHYR", "VISITYR", "VISITMO", "VISITDAY", "NACCUDSD", "NACCMMSE",
    "BOSTON", "LOGIMEM", "MEMUNITS", "MEMTIME", "DIGIF", "DIGIFLEN", "DIGIB", "DIGIBLEN"),
  methods = NULL,
  print_messages = F,
  with_diags = T
)

Arguments

dat

data set similar to the NACC data. For an example, see ?demo_data.

selected_cols

vector with columns to keep. If named, the names must correspond to NACC variable names and entries columns in the data set dat.

methods

NULL (default) or list of named entries specifying which model to use for standardizing cognitive scores. If NULL, defaults are applied for all variables in the dataset for which methods have been implemented (see NpsychBatteryNorms::default_methods)

print_messages

logical; should messages be printed? If TRUE (default), a message will be print with the methods used if methods is omitted (or NULL)

with_diags

logical (default TRUE); should diagnoses be included?

Examples

prepare_data(demo_data)
#>       RACE CDRGLOB raw_MOCATOTS MOCBTOTS raw_TRAILA TRAILARR TRAILALI
#>      <num>   <num>        <num>    <num>      <num>    <num>    <num>
#>   1:     2     0.5           24       -4        150       97       24
#>   2:     2     0.5           25       -4         26        0       24
#>   3:     1     0.5           16       -4         45        0       24
#>   4:     1     0.5           -4       -4         28        0       96
#>   5:     1     0.0           -4       -4         -4       -4       -4
#>  ---                                                                 
#> 361:     1     0.5           -4       -4         63       -4       -4
#> 362:     1     0.5           -4       -4         24        0       24
#> 363:     1     1.0           -4       -4         29        0       24
#> 364:     1     0.5           -4       -4         42        0       24
#> 365:     1     3.0           -4       -4         27        0       24
#>      raw_OTRAILA raw_OTRLARR raw_DIGFORCT raw_DIGFORSL raw_DIGBACCT
#>            <num>       <num>        <num>        <num>        <num>
#>   1:          -4          -4           97            8            7
#>   2:          -4          -4           12            6            5
#>   3:          -4          -4            6            3            4
#>   4:          -4          -4           -4           -4           -4
#>   5:          -4          -4           -4           -4           -4
#>  ---                                                               
#> 361:          -4          -4           -4           -4           -4
#> 362:          -4          -4           -4           -4           -4
#> 363:          -4          -4           -4           -4           -4
#> 364:          -4          -4           -4           -4           -4
#> 365:          -4          -4           -4           -4           -4
#>      raw_DIGBACLS raw_WAIS raw_MINTTOTS raw_ANIMALS raw_VEG raw_UDSVERTN
#>             <num>    <num>        <num>       <num>   <num>        <num>
#>   1:            4       -4           32          29      13           35
#>   2:            5       -4           28          15       7           14
#>   3:            5       -4           26           9      14           33
#>   4:           -4       96           -4          20       9           -4
#>   5:           -4       -4           -4          -4      -4           -4
#>  ---                                                                    
#> 361:           -4       48           -4          19      34           -4
#> 362:           -4       51           -4          18       7           -4
#> 363:           -4       98           -4          28       8           -4
#> 364:           -4       16           -4          31      10           -4
#> 365:           -4       96           -4          31      18           -4
#>      raw_UDSVERFC raw_UDSVERLC raw_UDSBENTC raw_UDSBENTD raw_CRAFTVRS
#>             <num>        <num>        <num>        <num>        <num>
#>   1:           19            9           10           11           21
#>   2:           17            8           16           12           38
#>   3:           12           11           16            7           12
#>   4:           -4           -4           -4           -4           -4
#>   5:           -4           -4           -4           -4           -4
#>  ---                                                                 
#> 361:           -4           -4           -4           -4           -4
#> 362:           -4           -4           -4           -4           -4
#> 363:           -4           -4           -4           -4           -4
#> 364:           -4           -4           -4           -4           -4
#> 365:           -4           -4           -4           -4           -4
#>      raw_CRAFTURS raw_CRAFTDVR raw_CRAFTDRE raw_REY1REC raw_REY2REC raw_REY3REC
#>             <num>        <num>        <num>       <num>       <num>       <num>
#>   1:            7           21            1          -4          -4          -4
#>   2:           15           16           13          -4          -4          -4
#>   3:           10           11           14          -4          -4          -4
#>   4:           -4           -4           -4          -4          -4          -4
#>   5:           -4           -4           -4          -4          -4          -4
#>  ---                                                                           
#> 361:           -4           -4           -4          -4          -4          -4
#> 362:           -4           -4           -4          -4          -4          -4
#> 363:           -4           -4           -4          -4          -4          -4
#> 364:           -4           -4           -4          -4          -4          -4
#> 365:           -4           -4           -4          -4          -4          -4
#>      raw_REY4REC raw_REY5REC raw_REY6REC raw_REYDREC REYTCOR raw_TRAILB
#>            <num>       <num>       <num>       <num>   <num>      <num>
#>   1:          -4          -4          -4          -4      -4        998
#>   2:          -4          -4          -4          -4      -4         69
#>   3:          -4          -4          -4          -4      -4         36
#>   4:          -4          -4          -4          -4      -4         75
#>   5:          -4          -4          -4          -4      -4         -4
#>  ---                                                                   
#> 361:          -4          -4          -4          -4      -4         76
#> 362:          -4          -4          -4          -4      -4        996
#> 363:          -4          -4          -4          -4      -4         70
#> 364:          -4          -4          -4          -4      -4        115
#> 365:          -4          -4          -4          -4      -4        101
#>      TRAILBLI TRAILBRR MOCACLOC MOCACLOH MOCACLON raw_OTRAILB raw_OTRLBRR
#>         <num>    <num>    <num>    <num>    <num>       <num>       <num>
#>   1:       21       97        1        0        0          -4          -4
#>   2:       24        1        1        0        1          -4          -4
#>   3:       24        1        1       98        1          -4          -4
#>   4:       24        0       -4       -4       -4          -4          -4
#>   5:       -4       -4       -4       -4       -4          -4          -4
#>  ---                                                                     
#> 361:       -4       -4       -4       -4       -4          -4          -4
#> 362:       24        1       -4       -4       -4          -4          -4
#> 363:       96       96       -4       -4       -4          -4          -4
#> 364:       24        0       -4       -4       -4          -4          -4
#> 365:       24        0       -4       -4       -4          -4          -4
#>      OTRLBLI NACCGDS CDRSUM UDSBENRS     NACCID   SEX  EDUC HANDED NACCAGE
#>        <num>   <num>  <num>    <num>     <char> <num> <num>  <num>   <num>
#>   1:      -4       3    0.0        1 NACC097622     2    16      2      72
#>   2:      -4       0    0.5        1 NACC097622     2    16      2      80
#>   3:      -4       4    0.0        1 NACC030397     1    13      2      77
#>   4:      -4       2    0.0       -4 NACC069324     1    14      2      82
#>   5:      -4      -4    3.5       -4 NACC069324     1    14      2      84
#>  ---                                                                      
#> 361:      -4       0    0.5       -4 NACC073678     1    16      1      95
#> 362:      -4       0    0.0       -4 NACC073678     1    16      1      91
#> 363:      -4       1    0.0       -4 NACC073678     1    16      1      77
#> 364:      -4       0    0.0       -4 NACC073678     1    16      1      87
#> 365:      -4       3    2.0       -4 NACC073678     1    16      1      70
#>      BIRTHYR VISITYR VISITMO VISITDAY NACCUDSD raw_NACCMMSE raw_BOSTON
#>        <num>   <num>   <num>    <num>    <num>        <num>      <num>
#>   1:    1935    2007       7       27        3           -4         -4
#>   2:    1935    2018       6        6        1           -4         -4
#>   3:    1943    2020       4       14        1           -4         -4
#>   4:    1925    2007      12       14        1           26         27
#>   5:    1925    2015       4       22        1           -4         -4
#>  ---                                                                  
#> 361:    1926    2021      10       17        3           29         26
#> 362:    1926    2023       8        3        1           27         28
#> 363:    1926    2009      10       12        1           30         29
#> 364:    1926    2022       3       25        4           24         27
#> 365:    1926    2011       4       21        1           30         26
#>      raw_LOGIMEM raw_MEMUNITS MEMTIME raw_DIGIF raw_DIGIFLEN raw_DIGIB
#>            <num>        <num>   <num>     <num>        <num>     <num>
#>   1:          -4           -4      -4        -4           -4        -4
#>   2:          -4           -4      -4        -4           -4        -4
#>   3:          -4           -4      -4        -4           -4        -4
#>   4:          20           97      -4         8            4         6
#>   5:          -4           -4      -4        -4           -4        -4
#>  ---                                                                  
#> 361:           2            0      18        98            8         8
#> 362:           9           13      14         9            6         6
#> 363:          13           15      15        10            6         6
#> 364:          96           18      22         6            8         6
#> 365:          97           15      20         8            4         6
#>      raw_DIGIBLEN ALCDEM_etiology ALCDEM_contribution ANXIET_etiology
#>             <num>           <num>               <num>           <num>
#>   1:           -4               0                   7               0
#>   2:           -4               0                   8               0
#>   3:           -4               0                   7               0
#>   4:            4               0                   7              -4
#>   5:           -4               0                   7              -4
#>  ---                                                                 
#> 361:            6               8                   8              -4
#> 362:            4               0                   7              -4
#> 363:           96               0                   7              -4
#> 364:            2               0                   8              -4
#> 365:            5               0                   7              -4
#>      ANXIET_contribution BIPOLDX_contribution BIPOLDX_etiology BRNINJ_etiology
#>                    <num>                <num>            <num>           <num>
#>   1:                   8                    7                0               0
#>   2:                   8                    8                0               0
#>   3:                   7                    7                0               0
#>   4:                  -4                   -4               -4               0
#>   5:                  -4                   -4               -4               0
#>  ---                                                                          
#> 361:                  -4                   -4               -4               0
#> 362:                  -4                   -4               -4               0
#> 363:                  -4                   -4               -4               0
#> 364:                  -4                   -4               -4               0
#> 365:                  -4                   -4               -4               0
#>      BRNINJ_contribution COGOTH_etiology COGOTH2_etiology COGOTH2_contribution
#>                    <num>          <char>           <char>                <num>
#>   1:                   7               0                0                    8
#>   2:                   7               0                0                    7
#>   3:                   7               0                0                    8
#>   4:                   8               0                0                    8
#>   5:                   8               0                0                    7
#>  ---                                                                          
#> 361:                   7               0               -4                    8
#> 362:                   7               0                0                    7
#> 363:                   8               0                0                    8
#> 364:                   7               0                0                    7
#> 365:                   8               0                0                    8
#>      COGOTH2X COGOTH3_etiology COGOTH3X COGOTH_contribution COGOTHX
#>        <char>           <char>   <char>               <num>  <char>
#>   1:                         0                            8        
#>   2:                         0                            8        
#>   3:                         0                            7        
#>   4:                         0                            7        
#>   5:                         0                            8        
#>  ---                                                               
#> 361:                        -4                            7        
#> 362:                         0                            7        
#> 363:                         0                            8        
#> 364:                         0                            7        
#> 365:                         0                            7        
#>      CORT_etiology CORT_contribution CVD_etiology CVD_contribution
#>              <num>             <num>        <num>            <num>
#>   1:             0                 7            0                3
#>   2:             1                 7            0                1
#>   3:             0                 8            0                7
#>   4:             0                 8           -4               -4
#>   5:             0                 8           -4               -4
#>  ---                                                              
#> 361:             0                 8           -4               -4
#> 362:             0                 7           -4               -4
#> 363:             0                 8           -4               -4
#> 364:             0                 8           -4               -4
#> 365:             0                 7           -4               -4
#>      DELIR_etiology DELIR_contribution DEMUN_etiology DEMUN_contribution
#>               <num>              <num>          <num>              <num>
#>   1:              0                  7             -4                 -4
#>   2:              0                  8             -4                 -4
#>   3:              0                  8             -4                 -4
#>   4:             -4                 -4              8                  8
#>   5:             -4                 -4              0                  8
#>  ---                                                                    
#> 361:             -4                 -4              8                  7
#> 362:             -4                 -4              8                  7
#> 363:             -4                 -4              8                  7
#> 364:             -4                 -4              0                  7
#> 365:             -4                 -4              8                  7
#>      DEP_etiology DEP_contribution DOWNS_etiology DOWNS_contribution
#>             <num>            <num>          <num>              <num>
#>   1:            1                8              0                  8
#>   2:            0                8              0                  7
#>   3:            0                8              0                  7
#>   4:            1                8              0                  7
#>   5:            0                3              0                  8
#>  ---                                                                
#> 361:            0                7              0                  7
#> 362:            1                7              0                  7
#> 363:            0                8              0                  8
#> 364:            0                8              0                  8
#> 365:            0                7              0                  8
#>      DYSILL_etiology DYSILL_contribution EPILEP_etiology EPILEP_contribution
#>                <num>               <num>           <num>               <num>
#>   1:               0                   8               0                   8
#>   2:               0                   8               0                   8
#>   3:               0                   8               0                   8
#>   4:               0                   7              -4                  -4
#>   5:               0                   2              -4                  -4
#>  ---                                                                        
#> 361:               0                   7              -4                  -4
#> 362:               0                   7              -4                  -4
#> 363:               0                   8              -4                  -4
#> 364:               0                   7              -4                  -4
#> 365:               0                   7              -4                  -4
#>      ESSTREM_contribution ESSTREM_etiology FTLDMO_etiology FTLDMO_contribution
#>                     <num>            <num>           <num>               <num>
#>   1:                    8                0               0                   7
#>   2:                    8                0               0                   8
#>   3:                    8                0               0                   8
#>   4:                   -4               -4              -4                  -4
#>   5:                   -4               -4              -4                  -4
#>  ---                                                                          
#> 361:                   -4               -4              -4                  -4
#> 362:                   -4               -4              -4                  -4
#> 363:                   -4               -4              -4                  -4
#> 364:                   -4               -4              -4                  -4
#> 365:                   -4               -4              -4                  -4
#>      FTLDNOS_contribution FTLDNOS_etiology HIV_etiology HIV_contribution
#>                     <num>            <num>        <num>            <num>
#>   1:                    7                0            0                7
#>   2:                    7                0            0                7
#>   3:                    7                1            0                7
#>   4:                   -4               -4           -4               -4
#>   5:                   -4               -4           -4               -4
#>  ---                                                                    
#> 361:                   -4               -4           -4               -4
#> 362:                   -4               -4           -4               -4
#> 363:                   -4               -4           -4               -4
#> 364:                   -4               -4           -4               -4
#> 365:                   -4               -4           -4               -4
#>      HUNT_etiology HUNT_contribution HYCEPH_etiology HYCEPH_contribution
#>              <num>             <num>           <num>               <num>
#>   1:             0                 7               0                   8
#>   2:             0                 8               0                   8
#>   3:             0                 8               0                   8
#>   4:             0                 8               0                   8
#>   5:             0                 8               0                   7
#>  ---                                                                    
#> 361:             0                 8               0                   8
#> 362:             0                 7               0                   7
#> 363:             0                 8               0                   7
#> 364:             0                 7               0                   7
#> 365:             0                 8               0                   8
#>      IMPSUB_etiology IMPSUB_contribution MEDS_etiology MEDS_contribution
#>                <num>               <num>         <num>             <num>
#>   1:               0                   8             0                 8
#>   2:               0                   8             1                 8
#>   3:               0                   8             0                 7
#>   4:              -4                  -4             0                 8
#>   5:              -4                  -4             0                 8
#>  ---                                                                    
#> 361:              -4                  -4             0                 8
#> 362:              -4                  -4             0                 8
#> 363:              -4                  -4             0                 8
#> 364:              -4                  -4             0                 7
#> 365:              -4                  -4             0                 7
#>      MSA_etiology MSA_contribution NACCALZD_etiology NACCALZD_contribution
#>             <num>            <num>             <num>                 <num>
#>   1:            0                7                 8                     1
#>   2:            0                7                 8                     8
#>   3:            0                7                 1                     7
#>   4:           -4               -4                 1                     1
#>   5:           -4               -4                 8                     1
#>  ---                                                                      
#> 361:           -4               -4                 1                     1
#> 362:           -4               -4                 0                     1
#> 363:           -4               -4                 0                     8
#> 364:           -4               -4                 1                     8
#> 365:           -4               -4                 0                     1
#>      NACCLBDE_etiology NACCLBDE_contribution NEOP_etiology NEOP_contribution
#>                  <num>                 <num>         <num>             <num>
#>   1:                 8                     7             0                 8
#>   2:                 0                     7             0                 8
#>   3:                 0                     8             0                 8
#>   4:                 0                     8             0                 8
#>   5:                 1                     8             0                 8
#>  ---                                                                        
#> 361:                 0                     7             0                 7
#> 362:                 8                     8             0                 8
#> 363:                 0                     8             0                 7
#> 364:                 8                     1             0                 7
#> 365:                 0                     7             0                 8
#>      OTHCOG_etiology OTHCOG_contribution OTHCOGX OTHPSY_etiology
#>               <char>               <num>  <char>          <char>
#>   1:               0                   7                       0
#>   2:               0                   8                       0
#>   3:               0                   1                       0
#>   4:              -4                  -4                       0
#>   5:              -4                  -4                       0
#>  ---                                                            
#> 361:              -4                  -4                       0
#> 362:              -4                  -4                       0
#> 363:              -4                  -4                       0
#> 364:              -4                  -4                       0
#> 365:              -4                  -4                       0
#>      OTHPSY_contribution OTHPSYX POSSAD_etiology POSSAD_contribution
#>                    <num>  <char>           <num>               <num>
#>   1:                   8                      -4                  -4
#>   2:                   8                      -4                  -4
#>   3:                   7                      -4                  -4
#>   4:                   7                       8                   7
#>   5:                   8                       8                   1
#>  ---                                                                
#> 361:                   7                       0                   7
#> 362:                   7                       0                   7
#> 363:                   7                       0                   7
#> 364:                   7                       0                   7
#> 365:                   7                       0                   7
#>      PPAPH_etiology PPAPH_contribution PRION_etiology PRION_contribution
#>               <num>              <num>          <num>              <num>
#>   1:             -4                 -4              0                  7
#>   2:             -4                 -4              0                  8
#>   3:             -4                 -4              0                  7
#>   4:              0                  7              0                  8
#>   5:              0                  7              0                  8
#>  ---                                                                    
#> 361:              0                  7              0                  7
#> 362:              8                  1              0                  7
#> 363:              8                  7              0                  7
#> 364:              0                  8              0                  8
#> 365:              0                  7              0                  8
#>      PROBAD_etiology PROBAD_contribution PSP_etiology PSP_contribution
#>                <num>               <num>        <num>            <num>
#>   1:              -4                  -4            0                7
#>   2:              -4                  -4            0                8
#>   3:              -4                  -4            0                7
#>   4:               1                   7            0                8
#>   5:               0                   8            0                8
#>  ---                                                                  
#> 361:               0                   1            0                8
#> 362:               1                   7            0                8
#> 363:               0                   1            0                7
#> 364:               8                   7            0                7
#> 365:               0                   7            0                7
#>      PTSDDX_etiology PTSDDX_contribution SCHIZOP_contribution SCHIZOP_etiology
#>                <num>               <num>                <num>            <num>
#>   1:               0                   8                    8                0
#>   2:               0                   7                    7                0
#>   3:               0                   7                    7                0
#>   4:              -4                  -4                   -4               -4
#>   5:              -4                  -4                   -4               -4
#>  ---                                                                          
#> 361:              -4                  -4                   -4               -4
#> 362:              -4                  -4                   -4               -4
#> 363:              -4                  -4                   -4               -4
#> 364:              -4                  -4                   -4               -4
#> 365:              -4                  -4                   -4               -4
#>      STROKE_etiology STROKE_contribution VASC_etiology VASC_contribution
#>                <num>               <num>         <num>             <num>
#>   1:              -4                  -4            -4                -4
#>   2:              -4                  -4            -4                -4
#>   3:              -4                  -4            -4                -4
#>   4:               0                   8             8                 7
#>   5:               0                   7             0                 7
#>  ---                                                                    
#> 361:               0                   8             0                 7
#> 362:               0                   7             8                 8
#> 363:               0                   8             0                 8
#> 364:               0                   7             8                 7
#> 365:               0                   8             8                 7
#>      VASCPS_etiology VASCPS_contribution BILLS TAXES SHOPPING GAMES STOVE
#>                <num>               <num> <num> <num>    <num> <num> <num>
#>   1:              -4                  -4     1     8        0     0     3
#>   2:              -4                  -4     0     0        0     3     0
#>   3:              -4                  -4     0     8        0     3     0
#>   4:               8                   7     0     0        0     0     8
#>   5:               0                   8     2     3        2     0     0
#>  ---                                                                     
#> 361:              -4                  -4     0     1        0     0     0
#> 362:               8                   7     1     0        2     0     0
#> 363:               0                   7     0     0        3     0     0
#> 364:               0                   8     0     0        0     0     0
#> 365:               0                   8     0     0        0     8     0
#>      MEALPREP EVENTS PAYATTN REMDATES TRAVEL REYFPOS  VISITDATE REYTOTAL
#>         <num>  <num>   <num>    <num>  <num>   <num>     <Date>    <num>
#>   1:        0      0       0        0      3      -4 2007-07-27       NA
#>   2:        2      0       1        0      0      -4 2018-06-06       NA
#>   3:        2      2       0        0      0      -4 2020-04-14       NA
#>   4:        0      2       2        0      3      -4 2007-12-14       NA
#>   5:        0      0       0        0      0      -4 2015-04-22       NA
#>  ---                                                                    
#> 361:        0      1       0        0      3      -4 2021-10-17       NA
#> 362:        0      0       0        0      0      -4 2023-08-03       NA
#> 363:        0      0       1        3      0      -4 2009-10-12       NA
#> 364:        0      0       2        0      0      -4 2022-03-25       NA
#> 365:        0      0       0        0      0      -4 2011-04-21       NA
#>      REYAREC   FAS MOCACLOCK std_MOCATOTS std_OTRAILA std_OTRAILB std_OTRLARR
#>        <num> <num>     <num>        <num>       <num>       <num>       <num>
#>   1:      NA     1         1   -0.9277594          NA          NA          NA
#>   2:      NA     3         2   -0.2776036          NA          NA          NA
#>   3:      NA     4        NA   -3.4495270          NA          NA          NA
#>   4:      NA     4        NA           NA          NA          NA          NA
#>   5:      NA     4        NA           NA          NA          NA          NA
#>  ---                                                                         
#> 361:      NA     2        NA           NA          NA          NA          NA
#> 362:      NA     3        NA           NA          NA          NA          NA
#> 363:      NA     1        NA           NA          NA          NA          NA
#> 364:      NA     2        NA           NA          NA          NA          NA
#> 365:      NA     0        NA           NA          NA          NA          NA
#>      std_OTRLBRR std_DIGFORCT std_DIGFORSL std_DIGBACCT std_DIGBACLS
#>            <num>        <num>        <num>        <num>        <num>
#>   1:          NA           NA    1.1206832  -0.03147211  -0.82148477
#>   2:          NA    1.8211550   -0.3507006  -0.82753134   0.08498118
#>   3:          NA   -0.8044175   -2.6489584  -1.14780633   0.22661649
#>   4:          NA           NA           NA           NA           NA
#>   5:          NA           NA           NA           NA           NA
#>  ---                                                                
#> 361:          NA           NA           NA           NA           NA
#> 362:          NA           NA           NA           NA           NA
#> 363:          NA           NA           NA           NA           NA
#> 364:          NA           NA           NA           NA           NA
#> 365:          NA           NA           NA           NA           NA
#>       std_TRAILA std_TRAILB std_WAIS std_MINTTOTS std_ANIMALS    std_VEG
#>            <num>      <num>    <num>        <num>       <num>      <num>
#>   1: -11.0849046         NA       NA     1.099705  1.45784924 -0.7332402
#>   2:   0.7708628  0.6822337       NA    -0.658333 -0.98799391 -2.0613359
#>   3:  -0.9183443  1.6670876       NA    -1.698443 -1.82732118  0.4728976
#>   4:   0.8012512  0.8038074       NA           NA  0.30496702 -0.7459338
#>   5:          NA         NA       NA           NA          NA         NA
#>  ---                                                                    
#> 361:  -2.0471774  1.0703400  67.8617           NA  0.26489000  5.5896501
#> 362:   1.4109716         NA  72.2232           NA -0.03893196 -1.2131069
#> 363:   0.3647542  0.5001896       NA           NA  1.47407089 -1.2534276
#> 364:  -0.4279473 -0.1969854  25.2136           NA  2.32904727 -0.5490590
#> 365:   0.2598351 -0.5326757       NA           NA  1.84888643  1.0909002
#>      std_UDSVERFC std_UDSVERLC std_UDSVERTN std_UDSBENTC std_UDSBENTD
#>             <num>        <num>        <num>        <num>        <num>
#>   1:    0.8547025   -1.2580645    0.6996069   -4.2418917  -0.01832598
#>   2:    0.5049611   -1.3971707   -1.7776575    0.3817703   0.55181552
#>   3:   -0.2391224   -0.2657585    0.9664268    0.5514459  -1.23021604
#>   4:           NA           NA           NA           NA           NA
#>   5:           NA           NA           NA           NA           NA
#>  ---                                                                 
#> 361:           NA           NA           NA           NA           NA
#> 362:           NA           NA           NA           NA           NA
#> 363:           NA           NA           NA           NA           NA
#> 364:           NA           NA           NA           NA           NA
#> 365:           NA           NA           NA           NA           NA
#>      std_CRAFTVRS std_CRAFTURS std_CRAFTDVR std_CRAFTDRE std_REY1REC
#>             <num>        <num>        <num>        <num>       <num>
#>   1:   -0.1844847    -2.354220    0.2729675   -3.4429287          NA
#>   2:    2.6346917    -0.162360   -0.3205618   -0.3363516          NA
#>   3:   -1.1502152    -1.125104   -0.8456548    0.1937968          NA
#>   4:           NA           NA           NA           NA          NA
#>   5:           NA           NA           NA           NA          NA
#>  ---                                                                
#> 361:           NA           NA           NA           NA          NA
#> 362:           NA           NA           NA           NA          NA
#> 363:           NA           NA           NA           NA          NA
#> 364:           NA           NA           NA           NA          NA
#> 365:           NA           NA           NA           NA          NA
#>      std_REY2REC std_REY3REC std_REY4REC std_REY5REC std_REY6REC std_REYDREC
#>            <num>       <num>       <num>       <num>       <num>       <num>
#>   1:          NA          NA          NA          NA          NA          NA
#>   2:          NA          NA          NA          NA          NA          NA
#>   3:          NA          NA          NA          NA          NA          NA
#>   4:          NA          NA          NA          NA          NA          NA
#>   5:          NA          NA          NA          NA          NA          NA
#>  ---                                                                        
#> 361:          NA          NA          NA          NA          NA          NA
#> 362:          NA          NA          NA          NA          NA          NA
#> 363:          NA          NA          NA          NA          NA          NA
#> 364:          NA          NA          NA          NA          NA          NA
#> 365:          NA          NA          NA          NA          NA          NA
#>      std_NACCMMSE   std_BOSTON std_LOGIMEM std_MEMUNITS   std_DIGIF
#>             <num>        <num>       <num>        <num>       <num>
#>   1:           NA           NA          NA           NA          NA
#>   2:           NA           NA          NA           NA          NA
#>   3:           NA           NA          NA           NA          NA
#>   4:   -2.2296213  0.337590539   1.7331771           NA -0.07704392
#>   5:           NA           NA          NA           NA          NA
#>  ---                                                               
#> 361:    0.1713071  0.001854443  -3.2150678  -3.14232917          NA
#> 362:   -1.5076140  0.601360535  -1.3617279   0.01243196  0.34992799
#> 363:    0.6855330  0.694919447  -0.3748577   0.42036667  0.70709810
#> 364:   -3.9934722  0.198510422          NA   1.26359556 -1.19784693
#> 365:    0.5717008 -0.427716669          NA   0.40327494 -0.36961735
#>      std_DIGIFLEN  std_DIGIB std_DIGIBLEN
#>             <num>      <num>        <num>
#>   1:           NA         NA           NA
#>   2:           NA         NA           NA
#>   3:           NA         NA           NA
#>   4:   -2.3535660 -0.2288524  -0.60812474
#>   5:           NA         NA           NA
#>  ---                                     
#> 361:    1.3151209  0.6736001   1.04116589
#> 362:   -0.5744443 -0.3300045  -0.69617429
#> 363:   -0.6814937 -0.4788952           NA
#> 364:    1.2539498 -0.3725447  -2.43351447
#> 365:   -2.5939980 -0.5533406  -0.07464274